5 features

+++ app/controllers/questions_controller.rb
@@ -1,6 +1,12 @@
 class QuestionsController < ApplicationController
+  before_action :authenticate_user!, except: [:index, :show]
+
+  expose :questions, -> { Question.all }
+  expose :question
+  expose :answers, from: :question
+  expose :answer, -> { Answer.new }
+
   def index
-    @questions = Question.all
   end
 
   def show
@@ -10,9 +16,9 @@ class QuestionsController < ApplicationController
   end
 
   def create
-    @question = Question.new(question_params)
-    if @question.save
-      redirect_to @question
+    question.user = current_user
+    if question.save
+      redirect_to question, notice: 'Your question successfully created.'
     else
       render :new
     end
@@ -23,24 +29,18 @@ class QuestionsController < ApplicationController
 
   def update
     if question.update(question_params)
-      redirect_to @question
+      redirect_to question
     else
       render :edit
     end
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    question.destroy if current_user.owner?(question)
+    redirect_to questions_path, notice: 'Your question has been deleted.'
   end
 
   private
-
-  def question
-    @question ||= params[:id] ? Question.find(params[:id]) : Question.new
-  end
-
-  helper_method :question
 
   def question_params
     params.require(:question).permit(:title, :body)
